requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
pandas>=2.1.0
numpy>=1.24.0

# AI/ML dependencies
openai>=1.3.0
anthropic>=0.7.0
langchain>=0.0.350
langchain-openai>=0.0.2
langchain-anthropic>=0.0.1

# Data processing
aiohttp>=3.9.0
asyncio-throttle>=1.0.2
python-dotenv>=1.0.0

# Database (optional)
sqlite3
sqlalchemy>=2.0.0

# Utilities
schedule>=1.2.0
click>=8.1.0
rich>=13.7.0
loguru>=0.7.0

# Development
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0
